requirements.txt
accelerate==0.0.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
arxiv==2.1.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
bitsandbytes==0.42.0
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
dataclasses-json==0.5.14
datasets==2.16.1
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
distro==1.9.0
dnspython==2.4.2
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.109.0
feedparser==6.0.10
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kubernetes==29.0.0
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.12
langchain-openai==0.0.2.post1
langsmith==0.0.83
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest_asyncio==1.5.9
ninja==1.11.1.1
numexpr==2.8.8
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.8.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pinecone-client==2.2.4
pip==23.3.1
platformdirs==4.1.0
posthog==3.3.2
prompt-toolkit==3.0.42
protobuf==4.25.2
psutil==5.9.0
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyMuPDF==1.23.15
PyMuPDFb==1.23.9
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pysbd==0.3.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
ragas==0.0.22
rank-bm25==0.2.2
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.1
scipy==1.11.4
setuptools==68.2.2
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.2
starlette==0.35.1
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
wheel==0.41.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
